# api-service/src/main/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: DDB API v3 - Records
  version: 3.0.0-dev
  contact:
    name: Michael Büchner
    email: m.buechner@dnb.de
    url: https://github.com/mbuechner/ddb-api-v3
  description: >-
    Standalone development contract containing only the Records API. It preserves
    record creation, replacement, deletion, validation, representation discovery,
    source and display access, and small batches. RDF record input is RDF/XML only;
    RDF output is negotiated through HTTP Accept across all production RDF formats
    registered by Apache Jena RIOT. Large bodies are streamed. Request bodies and
    response representations support gzip content coding.
servers:
  # Resolve API calls against the origin serving this contract. Swagger UI therefore works
  # unchanged on localhost, behind the gateway, and in every deployment environment.
  - url: /
security:
  - oauth2: [records:read]
...
paths:
  /3/records:
    post:
      tags: [Records]
      operationId: createRecord
      summary: Create a record
      description: >-
        Parses the complete RDF/XML record with Apache Jena RIOT and stores it only
        when it is valid RDF. The client supplies the globally unique record ID;
        the API never derives or replaces it. An existing ID returns 409.
        A 202 response links to the write-operation monitor.
      security:
        - oauth2: [records:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestContentEncoding'
        - $ref: '#/components/parameters/AcceptEncoding'
      requestBody:
        $ref: '#/components/requestBodies/RecordWriteRequest'
      responses:
        '201':
          description: Record has been created
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/DcatCatalogRecord'
        '202':
          description: Accepted; ordered publication continues at the Location monitor
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            Location:
              $ref: '#/components/headers/Location'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /3/records/{recordId}:
    get:
      tags: [Records]
      operationId: getRecord
      summary: Get record metadata
      description: >-
        Returns administrative metadata as one DCAT 3 JSON-LD catalog record. Tombstones remain addressable with
        the current ETag so an authorized client can restore them conditionally. They are identified by
        publicationStatus ending in /WITHDRAWN together with invalidatedAtTime; DELETED is only the internal
        lifecycle value because the EU Dataset Status vocabulary has no DELETED concept. Published records are public; restricted
        states require records:read. A permanently purged or unknown ID returns 404.
      security:
        - {}
        - oauth2: [records:read]
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
        - $ref: '#/components/parameters/AcceptEncoding'
      responses:
        '200':
          description: Current record state without RDF content
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/DcatCatalogRecord'
              examples:
                catalogRecord:
                  $ref: '#/components/examples/DcatRecordMetadata'
        '304':
          description: The current record state has not changed
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    head:
      tags: [Records]
      operationId: headRecord
      summary: Inspect record metadata headers
      description: >-
        Applies the same conditions and authorization as GET without returning the
        administrative metadata body.
      security:
        - {}
        - oauth2: [records:read]
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
        - $ref: '#/components/parameters/AcceptEncoding'
      responses:
        '200':
          description: Headers for the current record state
          headers:
            Content-Length:
              $ref: '#/components/headers/ContentLength'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '304':
          description: The current record state has not changed
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Last-Modified:
              $ref: '#/components/headers/LastModified'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      tags: [Records]
      operationId: putRecord
      summary: Update a record
      description: >-
        The complete RDF/XML is parsed with Apache Jena RIOT before persistence.
        The record must already exist and its current ETag is required in If-Match.
        The ID supplied in the metadata must equal the path ID and is never changed.
        A 202 response links to the monitor.
      security:
        - oauth2: [records:write]
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequiredIfMatch'
        - $ref: '#/components/parameters/RequestContentEncoding'
        - $ref: '#/components/parameters/AcceptEncoding'
      requestBody:
        $ref: '#/components/requestBodies/RecordWriteRequest'
      responses:
        '200':
          description: Record updated or left unchanged by an idempotent retry
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/DcatCatalogRecord'
        '202':
          description: Accepted; ordered publication continues at the Location monitor
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            Location:
              $ref: '#/components/headers/Location'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      tags: [Records]
      operationId: patchRecordLifecycle
      summary: Create or remove a record tombstone
      description: >-
        Changes only the record lifecycle state. publicationStatus=DELETED creates a
        tombstone without physically erasing retained content.
        publicationStatus=PUBLISHED restores the retained non-deleted content.
        Repeating the current target state is idempotent. Every
        transition requires the current state ETag.
      security:
        - oauth2: [records:write]
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequiredIfMatch'
        - $ref: '#/components/parameters/RequestContentEncoding'
        - $ref: '#/components/parameters/AcceptEncoding'
      requestBody:
        $ref: '#/components/requestBodies/RecordLifecyclePatchRequest'
      responses:
        '200':
          description: Lifecycle transition completed or target state already active
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/DcatCatalogRecord'
        '202':
          description: Accepted; ordered lifecycle publication continues at Location
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            Location:
              $ref: '#/components/headers/Location'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Records]
      operationId: deleteRecord
      summary: Permanently purge a record
      description: >-
        Irreversibly removes the record, its retained source content,
        its tombstone, and derived representations. This exceptional administrative
        operation is not a lifecycle transition and cannot be undone. The
        records:purge scope is issued only to administrative principals; possession
        of the scope is necessary but the server must also enforce its independent
        administrator policy, retention rules, and legal holds. Purging may complete
        asynchronously across storage, search, and cache systems.
      security:
        - oauth2: [records:purge]
      x-authorization-policy:
        principalType: ADMIN
        exceptionalAccess: true
        scopeIsNecessaryButNotSufficient: true
      parameters:
        - $ref: '#/components/parameters/RecordId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequiredIfMatch'
        - $ref: '#/components/parameters/PurgeReason'
        - $ref: '#/components/parameters/AcceptEncoding'
      responses:
        '202':
          description: Administrative purge accepted and continuing at Location
          headers:
            Vary:
              $ref: '#/components/headers/VaryAcceptEncoding'
            Content-Encoding:
              $ref: '#/components/headers/ContentEncoding'
            Location:
              $ref: '#/components/headers/Location'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriteOperation'
        '204':
          description: Record permanently purged, or identical purge retry completed
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.0 Client Credentials. Authentication is optional for operations that
        explicitly include the anonymous security alternative.
      flows:
        clientCredentials:
          tokenUrl: /oauth2/token
          scopes:
            records:read: Read records and the change feed
            records:write: Create and replace records
            records:purge: Permanently purge records; issued to administrators only
            records:validate: Validate RDF/XML without storing it
            imports:read: Read the client's import jobs
            imports:write: Create, start, and cancel import jobs
            exports:read: Read the client's export jobs
            exports:write: Create export jobs
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >-
        Canonical lowercase UUID scoped to client, method, and normalized URI.
        Identical retries are replayed for seven days; reuse with a different request
        fingerprint returns 409. UUIDv7 is recommended. Never encode sensitive data.
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 36
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
        example: 018f6c1a-7b2d-7f3e-a9c1-4b5d6e7f8091
...
    AcceptEncoding:
      name: Accept-Encoding
      in: header
      required: false
      description: >-
        Supported response content codings are gzip and identity. Quality values are
        honored. If neither coding is acceptable, the server returns 406.
      schema:
        type: string
        example: gzip, identity;q=0.5
    RequestContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: >-
        Content coding of the complete request body. gzip and identity are supported;
        identity is assumed when the header is absent. Limits apply after decoding.
      schema:
        type: string
        enum: [gzip, identity]
        default: identity
  headers:
...
    Location:
      required: true
      description: Created resource or asynchronous status monitor
      schema:
        type: string
        format: uri-reference
    ETag:
      required: true
      description: >-
        Entity tag for the selected media type and content coding. Strong tags differ
        between identity and gzip representations; weak tags may be shared only when
        their weak comparison semantics are sufficient.
      schema:
        type: string
...
    ContentEncoding:
      required: false
      description: >-
        Content coding applied to the response representation. The header is absent
        for identity. Content-Length, when present, is the encoded byte length.
      schema:
        type: string
        const: gzip
    VaryAcceptEncoding:
      required: true
      schema:
        type: string
        const: Accept, Accept-Encoding
...
    RequestId:
      required: true
      description: Correlation ID for this HTTP request
      schema:
        type: string
        minLength: 8
        maxLength: 128
    RetryAfter:
      description: >-
        Recommended retry time. A non-negative delay in seconds is preferred; an
        HTTP-date can be used when an absolute time is required.
      schema:
        # HTTP field values are textual on the wire. Modelling both lexical forms as a
        # string avoids an anonymous primitive oneOf that Java generators cannot represent.
        type: string
        minLength: 1
        examples:
          - '120'
          - 'Wed, 21 Oct 2015 07:28:00 GMT'
  requestBodies:
    RecordWriteRequest:
      required: true
      description: >-
        Canonical RDF/XML is required and the content part has the fixed media type
        application/rdf+xml; no other RDF input serialization is accepted. The outer
        request remains multipart/form-data so record metadata can be transmitted
        without custom headers. An optional non-RDF source part preserves the exact
        original XML. sourceMediaType, sourceFormat, sourceFormatUri, and
        sourceFormatVersion are then required in metadata. sourceProfile and
        sourceProfileUri are optional and must occur together. These conditional
        constraints are enforced by the application.
      content:
        multipart/form-data:
          schema:
            type: object
            additionalProperties: false
            required: [metadata, content]
            properties:
              metadata:
                $ref: '#/components/schemas/RecordWriteMetadata'
              content:
                type: string
                format: binary
                description: UTF-8 encoded RDF/XML document
              source:
                type: string
                format: binary
                description: >-
                  Optional exact original LIDO, EAD, METS, MODS, or other XML. This is
                  retained provenance, not an alternative RDF record input.
          encoding:
            metadata:
              contentType: application/json
            content:
              contentType: application/rdf+xml
            source:
              contentType: application/xml, application/mets+xml, application/mods+xml
...
  schemas:
...
    WriteOperation:
      type: object
      additionalProperties: false
      required: [operationId, status, createdAt]
      properties:
        operationId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        status:
          type: string
          enum:
            - STAGED
            - PENDING_PUBLICATION
            - PUBLISHED
            - PENDING_PURGE
            - PURGED
            - FAILED
        createdAt:
          type: string
          format: date-time
        recordId:
          anyOf:
            - $ref: '#/components/schemas/RecordIdValue'
            - type: 'null'
        problem:
          $ref: '#/components/schemas/Problem'
...
    DcatCatalogRecord:
      type: object
      additionalProperties: false
      description: >-
        The single administrative metadata representation for a record. It is a compacted JSON-LD
        1.1 document whose root is dcat:CatalogRecord. Its versioned external context uses only
        DCAT 3, Dublin Core, FOAF, ADMS, SKOS, PROV-O, EARL and EDM terms.
        DDB URLs identify resources and application profiles, but define no vocabulary terms.
      required:
        - '@context'
        - '@id'
        - '@type'
        - primaryTopic
        - createdAt
        - modifiedAt
      properties:
        '@context':
          type: string
          format: uri-reference
          description: Absolute IRI of the versioned JSON-LD 1.1 context document
        '@id':
          type: string
          format: uri-reference
          description: IRI identifying this dcat:CatalogRecord metadata entity
        '@type':
          type: string
          const: dcat:CatalogRecord
        primaryTopic:
          $ref: '#/components/schemas/DcatDataset'
        createdAt:
          type: string
          format: date-time
          description: dcterms:issued date of the catalog record
        modifiedAt:
          type: string
          format: date-time
          description: dcterms:modified date of the catalog record
...
  responses:
    BadRequest:
      description: Invalid request, parameter, or cursor
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Supplied credential is invalid, or a protected operation has none
      headers:
        WWW-Authenticate:
          schema:
            type: string
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Client is authenticated but not authorized
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotAcceptable:
      description: None of the requested response media types can be provided
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict with the current state or idempotency key
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ContentTooLarge:
      description: Request content exceeds the byte limit
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnsupportedMediaType:
      description: Request media type is not supported
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...